==> 01_data_loading_cleaning.ipynb <==
# Import Libraries
import pandas as pd
import numpy as np

# Show all columns when displaying DataFrames
pd.set_option('display.max_columns', None)


# Load CSV files
orders = pd.read_csv("olist_orders_dataset.csv")
customers = pd.read_csv("olist_customers_dataset.csv")
order_items = pd.read_csv("olist_order_items_dataset.csv")
order_payments = pd.read_csv("olist_order_payments_dataset.csv")

# Print dataset shapes
print("Orders:", orders.shape)
print("Customers:", customers.shape)
print("Order Items:", order_items.shape)
print("Payments:", order_payments.shape)


# Preview the Data
print("Orders:", orders.head())
print("Customers:", customers.head())
print("Order Items:", order_items.head())
print("Payments:", order_payments.head())

# Merge the data
## Merge orders + customers
orders_customers = pd.merge(orders, customers, on='customer_id', how='inner')
orders_customers.head()

## Merge orders_customers + order_payments
orders_customers_payments = pd.merge(orders_customers, order_payments, on='order_id', how='left')
orders_customers_payments.head()

## Merge orders_customers_payments + order_items
orders_full_data = pd.merge(orders_customers_payments, order_items, on='order_id', how='left')
orders_full_data.head()

### Save the Merged Data
orders_full_data.to_csv('merged_orders_dataset.csv', index=False)

# Clean the Merged Data (Order_full_data)
## How large the data is
orders_full_data.shape
## Check for missing values
orders_full_data.isnull().sum().sort_values(ascending=False)

## Drop the rows with missing price
orders_full_data = orders_full_data.dropna(subset=['price'])


## Drop the 3 rows with missing payment info
orders_full_data = orders_full_data.dropna(subset=['payment_value'])


## Check for missing values again
orders_full_data.isnull().sum().sort_values(ascending=False)


## Drop the rows with missing approval date
orders_full_data = orders_full_data.dropna(subset=['order_approved_at'])


## Check for missing values again
orders_full_data.isnull().sum().sort_values(ascending=False)

# Save the cleaned data
orders_full_data.to_csv("orders_full_data_cleaned.csv", index=False)
