==> 03_feature_engineering.ipynb <==
# Import the labeled Data
import pandas as pd
df = pd.read_csv("orders_labeled.csv")

# Preview the data
df.head()

# Total number of orders per customer # Feature 1
orders_per_customer = df.groupby("customer_unique_id")["order_id"].nunique().reset_index()
orders_per_customer.columns = ["customer_unique_id", "num_orders"]

# Total money spent by each customer # Feature 2
payment_per_customer = df.groupby("customer_unique_id")["payment_value"].sum().reset_index()
payment_per_customer.columns = ["customer_unique_id", "total_payment"]

# Average payment per order # Feature 3
avg_payment = df.groupby("customer_unique_id")["payment_value"].mean().reset_index()
avg_payment.columns = ["customer_unique_id", "avg_payment"]

# Days since last purchase # feature 4
## Convert timestamp
df["order_purchase_timestamp"] = pd.to_datetime(df["order_purchase_timestamp"])

## Last purchase date
last_purchase_date = df.groupby("customer_unique_id")["order_purchase_timestamp"].max().reset_index()
last_purchase_date.columns = ["customer_unique_id", "last_purchase_date"]

## Reference date = latest date in the dataset
reference_date = df["order_purchase_timestamp"].max()

## Days since last purchase
last_purchase_date["days_since_last_purchase"] = (reference_date - last_purchase_date["last_purchase_date"]).dt.days

# Merge all features into one table
features = orders_per_customer.merge(payment_per_customer, on="customer_unique_id")
features = features.merge(avg_payment, on="customer_unique_id")
features = features.merge(last_purchase_date[["customer_unique_id", "days_since_last_purchase"]], on="customer_unique_id")
features = features.merge(df[["customer_unique_id", "churn"]].drop_duplicates(), on="customer_unique_id")

# Convert it into a CVS File
features.to_csv("churn_features.csv", index=False)

# Preview the result
features.head()
